+++ launchdarkly_api/api/users_api.py
@@ -160,7 +160,7 @@ class UsersApi(object):
         :param str q: Search query.
         :param int limit: Pagination limit.
         :param int offset: Specifies the first item to return in the collection.
-        :param float after: A timestamp filter, expressed as a Unix epoch time in milliseconds. All entries returned will have occured after this timestamp.
+        :param int after: A timestamp filter, expressed as a Unix epoch time in milliseconds. All entries returned will have occured after this timestamp.
         :return: Users
                  If the method is called asynchronously,
                  returns the request thread.
@@ -186,7 +186,7 @@ class UsersApi(object):
         :param str q: Search query.
         :param int limit: Pagination limit.
         :param int offset: Specifies the first item to return in the collection.
-        :param float after: A timestamp filter, expressed as a Unix epoch time in milliseconds. All entries returned will have occured after this timestamp.
+        :param int after: A timestamp filter, expressed as a Unix epoch time in milliseconds. All entries returned will have occured after this timestamp.
         :return: Users
                  If the method is called asynchronously,
                  returns the request thread.
